Extended event log, support for file operations (open, close, read, write)
An extended event log helps comparing recordings of different runs of
the same process.
For this reason the log is output in json format and the comparison can
be done by another script that understands the format and knows what to
look for.

There is a long list of TODOs at the top of the python script before
this can be considered complete.