Add a new operation to replace reverse-stepping.
This should work more like the LLM expects, in that it'll reverse step into a
function on the current line.  It will work regardless of whether we used
`reverse-next` to get to that line (where `reverse-step` would not work) or used
`last` on the returned value (in which case `reverse-step`, confusingly, did
work).