Initial support for Amp coding agent.
This enables the Amp coding agent to act as a backend to the "explain" command.

Note that this implementation currently discards existing user configuration to
"amp" in order to apply an MCP server configuration.

Select the coding agent using the `--agent` (`-a`) flag to `explain` or the `EXPLAIN_AGENT`
environment variable.  Valid values for these are "claude" and "amp".

It is not possible to switch coding agent once one has been invoked during a session.