explain: Use Rich for formatting output.
The Python Rich library allows us to provide better quality console output. This
uses the version of Rich already shipped with UDB so there's no need to add the
package to the manifest entry.

This solves persistent formatting weirdnesses in the old scheme and adds a bit
more in the way of pretty console effects.

Using `force_terminal=True` enables us to use fancy formatting with Rich getting
confused by GDB's version of `sys.stdout` / `sys.stderr` but may prevent some
auto-detection behaviours working so we will want to revisit it in future.