Explain: Enable backtracing past main to handle clean process exit.
When a process exits cleanly, the `main()` stack frame is popped, which was
making it impossible for `explain` to step back into it (even if it could
reverse to the line where `main()` had been called, reverse stepping into a line
requires the ability to finish out of the function call of interest).

Enabling backtracing past main will make the C runtime parts of the backtrace
visible and also allows the desired finishing behaviour.

Note that for this to enable navigation of a cleanly exited program you will
probably also need debug info / source available for the C runtime so the LLM
can understand where it is.