Explain: Add current position within bookmarks.
Enable the LLM to reason more effectively about which bookmarks it needs to use
to navigate time by providing a clear indication of where it is currently.

This is based on the observation that, when it realises it has stepped too far
back, the LLM will often jump back to a bookmark it has set.  This additional
output is intended to make it easier to jump to a relevant point.