Explain: Simplify output from last-value tool.
Instead of supplying `last` console output, this provides a simplified version.
This is based on the observation that providing two values (before/after to the
LLM is potentially distracting when it's investigating where a particular value
came from).  Sometimes the LLM would focus on the earlier value from history as
being the result of the assignment.

All the agent really needs to know is that it's got back to just after the
assignment.  It can step back and discover the previous value if that's
relevant (which it won't always be).