Explain: Better feedback from "last" command.
Refuse to work on expressions that don't have an in-memory address, to avoid
poor performance.

Report explicitly when there's no change in value and only try to position the
debugger after the change if there actually was one (this avoids it running
unbounded amounts of forward execution if there was no change seen).