Explain: only stop on unhandled exceptions during a function call.
We shouldn't stop on throws that are caught within the execution of a function
when stepping back into it.  We only need to stop if we threw the exception out
instead of returning normally.

To handle this better, we use the `uncaught_exceptions` field in this thread's
C++ global state to determine if an unhandled exception was thrown, then (if so)
reverse back to the throw site.

The Itanium C++ ABI documentation is good to help understand the functions and
data structures involved here:
https://itanium-cxx-abi.github.io/cxx-abi/abi-eh.html