Explain: Handle alternate spelling of "uncaughtExceptions".
The symbol for this counter appears to have a couple of spellings that come up
in practice in different programs.

The field is still in the same place in the structure returned by
`__cxa_get_globals()` so in future we might want to just look at it by offset,
rather than name.  For now this will do.